fix(infra): give the canonical-key function a home that carries no native deps
Refs #142.

CI caught what I could not see locally. My "one derivation" import added an edge
from `current-instance.ts` — deliberately dependency-free, because the gateway,
the anatomy page and the Neighbor agent all read it — into `gateway-lock.ts`,
which `require`s `better-sqlite3` at module scope for the runtime lock.

The Dashboard UI suite imports `GatewayServer`, which knows which rappter it is,
so on a runner without that native module it died:

    Error: Cannot find module 'better-sqlite3'
    Require stack: typescript/src/infra/gateway-lock.ts
    Test Files  1 failed | 7 passed (8)

`npm test` in `typescript/` does not run `ui/`; only CI does. I had run the main
suite, tsc and lint and seen green.

`canonicalInstanceKey` now lives in `infra/instance-key.ts`, a leaf with no
imports at all, and `gateway-lock` re-exports it so every existing importer is
untouched. One implementation, lighter home — not a second derivation, which
would be the very defect #142 is about.

typescript: 4390 passed. ui: 8 files passed.
